pyproject.toml
[project]
name = "groq-site-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
    "groq>=0.30.0",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
